refactor(cli): convert to subcommand dynamic loading
