+++ semantic_release/__main__.py
@@ -1,6 +1,6 @@
 import sys
 
-from semantic_release.cli import main
+from semantic_release.cli.commands.main import main
 
 if __name__ == "__main__":
     main(args=sys.argv[1:])
